Split make_sortable into a utility module
